+++ app/src/main/java/com/konditsky/playlistmaker/data/MediaPlayerManager.kt
@@ -1,0 +1,32 @@
+package com.konditsky.playlistmaker.data
+
+import android.media.MediaPlayer
+import com.konditsky.playlistmaker.domain.TrackPlayer
+
+class MediaPlayerManager : TrackPlayer {
+    private val mediaPlayer = MediaPlayer()
+
+    override fun prepare(url: String) {
+        mediaPlayer.apply {
+            reset()
+            setDataSource(url)
+            prepareAsync()
+        }
+    }
+
+    override fun play() {
+        mediaPlayer.start()
+    }
+
+    override fun pause() {
+        mediaPlayer.pause()
+    }
+
+    override fun release() {
+        mediaPlayer.release()
+    }
+
+    override fun isPlaying(): Boolean {
+        return mediaPlayer.isPlaying
+    }
+}
