+++ app/src/main/java/com/konditsky/playlistmaker/domain/ImageLoader.kt
@@ -1,0 +1,7 @@
+package com.konditsky.playlistmaker.domain
+
+import android.widget.ImageView
+
+interface ImageLoader {
+    fun loadImage(url: String, into: ImageView)
+}
