+++ app/src/main/java/com/konditsky/playlistmaker/domain/TrackPlayer.kt
@@ -1,0 +1,9 @@
+package com.konditsky.playlistmaker.domain
+
+interface TrackPlayer {
+    fun prepare(url: String)
+    fun play()
+    fun pause()
+    fun release()
+    fun isPlaying(): Boolean
+}
