+++ app/src/main/java/com/konditsky/playlistmaker/presentation/MainActivity.kt
@@ -1,13 +1,10 @@
-package com.konditsky.playlistmaker
+package com.konditsky.playlistmaker.presentation
 
-import android.content.Context
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import android.view.View
 import android.widget.Button
-import android.widget.Toast
-import androidx.appcompat.app.AppCompatDelegate
+import com.konditsky.playlistmaker.R
 
 class MainActivity : AppCompatActivity() {
