+++ app/src/main/java/com/konditsky/playlistmaker/presentation/MediaActivity.kt
@@ -1,7 +1,8 @@
-package com.konditsky.playlistmaker
+package com.konditsky.playlistmaker.presentation
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import com.konditsky.playlistmaker.R
 
 class MediaActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
