+++ app/src/main/java/com/konditsky/playlistmaker/presentation/SettingsActivity.kt
@@ -1,15 +1,15 @@
-package com.konditsky.playlistmaker
+package com.konditsky.playlistmaker.presentation
 
 import android.content.Context
 import android.content.Intent
 import android.net.Uri
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import android.widget.Button
 import android.widget.ImageButton
 import android.widget.ImageView
 import android.widget.Switch
 import androidx.appcompat.app.AppCompatDelegate
+import com.konditsky.playlistmaker.R
 
 class SettingsActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
