+++ app/src/main/java/com/konditsky/playlistmaker/presentation/TrackAdapter.kt
@@ -1,9 +1,11 @@
-package com.konditsky.playlistmaker
+package com.konditsky.playlistmaker.presentation
 
 import android.util.Log
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
+import com.konditsky.playlistmaker.R
+import com.konditsky.playlistmaker.domain.Track
 
 class TrackAdapter(
     private var trackList: ArrayList<Track>,
