+++ app/src/main/java/com/konditsky/playlistmaker/presentation/TrackViewHolder.kt
@@ -1,4 +1,4 @@
-package com.konditsky.playlistmaker
+package com.konditsky.playlistmaker.presentation
 
 import android.view.View
 import android.widget.ImageView
@@ -8,6 +8,8 @@ import com.bumptech.glide.Glide
 import com.bumptech.glide.load.resource.bitmap.CenterCrop
 import com.bumptech.glide.load.resource.bitmap.RoundedCorners
 import com.bumptech.glide.request.RequestOptions
+import com.konditsky.playlistmaker.R
+import com.konditsky.playlistmaker.domain.Track
 
 class TrackViewHolder(view: View) : RecyclerView.ViewHolder(view) {
     private val trackNameTextView: TextView = view.findViewById(R.id.textTrackName)
