+++ app/src/main/res/layout/activity_media.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MediaActivity">
+    tools:context=".presentation.MediaActivity">
 
 
 </androidx.constraintlayout.widget.ConstraintLayout>
