+++ app/src/main/res/layout/activity_search.xml
@@ -5,7 +5,7 @@ style="@style/SettingsActivityStyle"
 xmlns:tools="http://schemas.android.com/tools"
 android:layout_width="match_parent"
 android:layout_height="match_parent"
-tools:context=".SearchActivity"
+tools:context=".presentation.SearchActivity"
 android:orientation="vertical">
